Split out types (DH-18428-2)
