+++ e2e-testing/src/util/index.ts
@@ -2,3 +2,4 @@ export * from './lifeCycle';
 export * from './locators';
 export * from './paths';
 export * from './testUtils';
+export type * from './types';
